Core FF: auxiliary classes from FFPolygon moved to extra files
